+++ examples/guest-link/register.rs
@@ -24,7 +24,7 @@ cfg_if::cfg_if! {
                 .join("../target/wasm32-wasip2/debug/examples");
 
             let manifest = Manifest::new()
-                .dispatch(["omnia:link/echo"])
+                .plugins(["omnia:link/echo"])
                 .guest(GuestEntry::new(
                     "responder",
                     artifacts.join("guest_link_responder_wasm.wasm"),